Add DataSource - This is an interface that every DataSource (Local or Remote) should implement.
Add Reponse<T> - This is a helper class that can be used to return results from async functions. It holds an object value or an error.

Add CaasDTO - This class is a data transfer object that will be used to hold all information received from the CAAS api.

Add CaasAPI - This class implements the Datasource interface and is responsable to interact with the remote service CAAS api.

Add CaasServiceTest - Test class to check if a request to CAAS service is returning the correct response.