CataasAPI - Replace the http library for the Retrofit library on Cataas API implementation.
CataasDTO - Added the JsonSerializable annotation and two methods fromJson and toJson, these methods are used in the serialization/deserialization process when a request is sent to CataasAPI.

CatRepository - Changed the way CataasAPI is called when a random cat image is requested.

Injector - Added an instance of Dio class to the creation of CataasAPI object.

cataas_service_test - Updated the test cases to reflect the changes of CataasAPI, also added more tests to check the response for different combinations of query and path parameters of the api.

catgif_repository_test - Updated the test cases to reflect the changes of CatRepository.